Prevent message box blocking subsequent notifications
Script updated so that the message box no longer waits for the user
to press the OK button, as that currently prevents it from displaying
further notifications about potential signups, and prevents it from
firing the Pushover request as well. Instead  we return the thread
handle and the caller function can decide to wait for the user or not.
Practically, this means waiting for the user to press the OK button
is only for when the thread limit has been reached message is displayed.

In addition, script updated to no longer exit when thread limit is
reached and instead will just no longer check the thread reply count.